Extract cluster configuration to kubernetes package
