+++ pkg/combustion/kubernetes_test.go
@@ -312,263 +312,3 @@ func TestConfigureKubernetes_SuccessfulMultiNodeRKE2Cluster(t *testing.T) {
 	assert.Equal(t, "https://192.168.122.100:9345", configContents["server"])
 	assert.Equal(t, []any{"192-168-122-100.sslip.io", "192.168.122.100", "api.cluster01.hosted.on.edge.suse.com"}, configContents["tls-san"])
 }
-
-func TestExtractCNI(t *testing.T) {
-	tests := map[string]struct {
-		input                 map[string]any
-		expectedCNI           string
-		expectedMultusEnabled bool
-		expectedErr           string
-	}{
-		"CNI not configured": {
-			input:       map[string]any{},
-			expectedErr: "invalid cni: <nil>",
-		},
-		"Empty CNI string": {
-			input: map[string]any{
-				"cni": "",
-			},
-			expectedErr: "cni not configured",
-		},
-		"Empty CNI list": {
-			input: map[string]any{
-				"cni": []string{},
-			},
-			expectedErr: "invalid cni value: []",
-		},
-		"Multiple CNI list": {
-			input: map[string]any{
-				"cni": []string{"canal", "calico", "cilium"},
-			},
-			expectedErr: "invalid cni value: [canal calico cilium]",
-		},
-		"Valid CNI string": {
-			input: map[string]any{
-				"cni": "calico",
-			},
-			expectedCNI: "calico",
-		},
-		"Valid CNI list": {
-			input: map[string]any{
-				"cni": []string{"calico"},
-			},
-			expectedCNI: "calico",
-		},
-		"Valid CNI string with multus": {
-			input: map[string]any{
-				"cni": "multus, calico",
-			},
-			expectedCNI:           "calico",
-			expectedMultusEnabled: true,
-		},
-		"Valid CNI list with multus": {
-			input: map[string]any{
-				"cni": []string{"multus", "calico"},
-			},
-			expectedCNI:           "calico",
-			expectedMultusEnabled: true,
-		},
-		"Invalid standalone multus": {
-			input: map[string]any{
-				"cni": "multus",
-			},
-			expectedErr: "multus must be used alongside another primary cni selection",
-		},
-		"Invalid standalone multus list": {
-			input: map[string]any{
-				"cni": []string{"multus"},
-			},
-			expectedErr: "multus must be used alongside another primary cni selection",
-		},
-		"Valid CNI with invalid multus placement": {
-			input: map[string]any{
-				"cni": "cilium, multus",
-			},
-			expectedErr: "multiple cni values are only allowed if multus is the first one",
-		},
-		"Valid CNI list with invalid multus placement": {
-			input: map[string]any{
-				"cni": []string{"cilium", "multus"},
-			},
-			expectedErr: "multiple cni values are only allowed if multus is the first one",
-		},
-		"Invalid CNI list": {
-			input: map[string]any{
-				"cni": []any{"cilium", 6},
-			},
-			expectedErr: "invalid cni value: 6",
-		},
-		"Invalid CNI format": {
-			input: map[string]any{
-				"cni": 6,
-			},
-			expectedErr: "invalid cni: 6",
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			b, err := yaml.Marshal(test.input)
-			require.NoError(t, err)
-
-			var config map[string]any
-			require.NoError(t, yaml.Unmarshal(b, &config))
-
-			cni, multusEnabled, err := extractCNI(config)
-
-			if test.expectedErr != "" {
-				require.Error(t, err)
-				assert.EqualError(t, err, test.expectedErr)
-				assert.False(t, multusEnabled)
-				assert.Empty(t, cni)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, test.expectedCNI, cni)
-				assert.Equal(t, test.expectedMultusEnabled, multusEnabled)
-			}
-		})
-	}
-}
-
-func TestFindKubernetesInitialiserNode(t *testing.T) {
-	tests := []struct {
-		name         string
-		nodes        []image.Node
-		expectedNode string
-	}{
-		{
-			name:         "Empty list of nodes",
-			expectedNode: "",
-		},
-		{
-			name: "Agent list",
-			nodes: []image.Node{
-				{
-					Hostname: "host1",
-					Type:     "agent",
-				},
-				{
-					Hostname: "host2",
-					Type:     "agent",
-				},
-			},
-
-			expectedNode: "",
-		},
-		{
-			name: "Server node labeled as initialiser",
-			nodes: []image.Node{
-				{
-					Hostname: "host1",
-					Type:     "agent",
-				},
-				{
-					Hostname: "host2",
-					Type:     "server",
-				},
-				{
-					Hostname:    "host3",
-					Type:        "server",
-					Initialiser: true,
-				},
-			},
-			expectedNode: "host3",
-		},
-		{
-			name: "Initialiser as first server node in list",
-			nodes: []image.Node{
-				{
-					Hostname: "host1",
-					Type:     "agent",
-				},
-				{
-					Hostname: "host2",
-					Type:     "server",
-				},
-				{
-					Hostname: "host3",
-					Type:     "server",
-				},
-			},
-			expectedNode: "host2",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			kubernetes := &image.Kubernetes{Nodes: test.nodes}
-			assert.Equal(t, test.expectedNode, findKubernetesInitialiserNode(kubernetes))
-		})
-	}
-}
-
-func TestAppendClusterTLSSAN(t *testing.T) {
-	tests := []struct {
-		name           string
-		config         map[string]any
-		apiHost        string
-		expectedTLSSAN any
-	}{
-		{
-			name:           "Empty TLS SAN",
-			config:         map[string]any{},
-			apiHost:        "",
-			expectedTLSSAN: nil,
-		},
-		{
-			name:           "Missing TLS SAN",
-			config:         map[string]any{},
-			apiHost:        "api.cluster01.hosted.on.edge.suse.com",
-			expectedTLSSAN: []string{"api.cluster01.hosted.on.edge.suse.com"},
-		},
-		{
-			name: "Invalid TLS SAN",
-			config: map[string]any{
-				"tls-san": 5,
-			},
-			apiHost:        "api.cluster01.hosted.on.edge.suse.com",
-			expectedTLSSAN: []string{"api.cluster01.hosted.on.edge.suse.com"},
-		},
-		{
-			name: "Existing TLS SAN string",
-			config: map[string]any{
-				"tls-san": "api.edge1.com, api.edge2.com",
-			},
-			apiHost:        "api.cluster01.hosted.on.edge.suse.com",
-			expectedTLSSAN: []string{"api.edge1.com", "api.edge2.com", "api.cluster01.hosted.on.edge.suse.com"},
-		},
-		{
-			name: "Existing TLS SAN string list",
-			config: map[string]any{
-				"tls-san": []string{"api.edge1.com", "api.edge2.com"},
-			},
-			apiHost:        "api.cluster01.hosted.on.edge.suse.com",
-			expectedTLSSAN: []string{"api.edge1.com", "api.edge2.com", "api.cluster01.hosted.on.edge.suse.com"},
-		},
-		{
-			name: "Existing TLS SAN list",
-			config: map[string]any{
-				"tls-san": []any{"api.edge1.com", "api.edge2.com"},
-			},
-			apiHost:        "api.cluster01.hosted.on.edge.suse.com",
-			expectedTLSSAN: []any{"api.edge1.com", "api.edge2.com", "api.cluster01.hosted.on.edge.suse.com"},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			appendClusterTLSSAN(test.config, test.apiHost)
-			assert.Equal(t, test.expectedTLSSAN, test.config["tls-san"])
-		})
-	}
-}
-
-func TestSetClusterAPIAddress(t *testing.T) {
-	config := map[string]any{}
-
-	setClusterAPIAddress(config, "")
-	assert.NotContains(t, config, "server")
-
-	setClusterAPIAddress(config, "192.168.122.50")
-	assert.Equal(t, "https://192.168.122.50:9345", config["server"])
-}
