+++ pkg/combustion/templates/15-rke2-single-node-installer.sh.tpl
@@ -3,23 +3,23 @@ set -euo pipefail
 
 mount /var
 mkdir -p /var/lib/rancher/rke2/agent/images/
-cp {{ .ImagesPath }}/* /var/lib/rancher/rke2/agent/images/
+cp {{ .imagesPath }}/* /var/lib/rancher/rke2/agent/images/
 
-{{- if .VIPManifest }}
+{{- if .vipManifest }}
 mkdir -p /var/lib/rancher/rke2/server/manifests/
-cp {{ .VIPManifest }} /var/lib/rancher/rke2/server/manifests/{{ .VIPManifest }}
+cp {{ .vipManifest }} /var/lib/rancher/rke2/server/manifests/{{ .vipManifest }}
 {{- end }}
 umount /var
 
-{{- if and .Network.APIVIP .Network.APIHost }}
-echo "{{ .Network.APIVIP }} {{ .Network.APIHost }}" >> /etc/hosts
+{{- if and .apiVIP .apiHost }}
+echo "{{ .apiVIP }} {{ .apiHost }}" >> /etc/hosts
 {{- end }}
 
 mkdir -p /etc/rancher/rke2/
-cp {{ .ConfigFile }} /etc/rancher/rke2/config.yaml
+cp {{ .configFile }} /etc/rancher/rke2/config.yaml
 
 export INSTALL_RKE2_TAR_PREFIX=/opt/rke2
-export INSTALL_RKE2_ARTIFACT_PATH={{ .InstallPath }}
+export INSTALL_RKE2_ARTIFACT_PATH={{ .installPath }}
 
 ./rke2_installer.sh
