+++ pkg/kubernetes/testdata/agent.yaml
@@ -1,0 +1,3 @@
+cni: calico
+token: totally-not-generated-one
+debug: true
