+++ pkg/kubernetes/testdata/server.yaml
@@ -1,0 +1,3 @@
+cni: calico
+token: totally-not-generated-one
+selinux: true
